Release nxtvepg-0.7.2 (2002-May-20)
Changes between 0.7.1 and 0.7.2 (as per file CHANGES):

- Windows: Enabled TV application interaction feature and added documentation
  for the demo and debugging tools tvsim, vbirec and vbiplay.
- Windows: Added option "Hide on minimize", i.e. when the main window is
  minimized, it's removed from the taskbar and an icon is installed in the
  system tray with a small popup menu.
- Windows TV app. interaction: VPS/PDC is now evaluated if EPG acquisition
  is disabled (or -in the future- if EPG acq. uses a different TV card)
- Windows TV card input dialog: warn if TV app. was changed but config not
  loaded; same for TV app. interaction dialog if "Test" button was not used.
- Windows Bt8x8 driver DSdrv: improved robustness by attempting uninstall
  and re-install of the service if driver start fails (with certain types
  of errors)
- Windows TV card input dialog: list only as many TV cards as are available
  (do a scan of all TV cards upon first Bt8x8 driver start; requires that
  the driver was started at least once since program start)
- Windows daemon port finished: option -daemon is now available (note that
  the daemon currently does not support TV app interaction)

- bugfix EPG scan (Linux only): after an EPG scan acquisition could not be
  enabled because the VBI device was allegedly busy. The cause was that an
  acquisition thread kept running after the EPG scan.
- the wheel mouse now scrolls only 5 lines instead of 1 page and the cursor
  remains on the previously selected item.
- improved automatic mode selection for internal/remote acquisition during
  startup: wait until an connection is completely established until changing
  the default mode to remote acq in the rc/ini file (esp. when connecting
  via TCP/IP)
- EPG scan is no longer prevented with "cannot start scan while in
  client/server mode" unless currently connected to an acq daemon.
- EPG scan: reset 45-seconds timer when EPG is found to be transmitted on a
  non-default teletext page nunmber. This should fix the problem that Canal+
  is reliably detected only when in "slow" mode.
- removed the last remaints of long integer usage to improve 64-bit CPU
  compatibility.
- bugfix: crash when there was an .xawtv file with zero channel definitions
- bugfix: the EPG scan was not stopped when the program was terminated by a
  signal or logging out, causing a crash.
- bugfix: crashed with SEGV when the acq timescales were open when the EPG
  scan was started
- bugfixes in timescales: network highlighting and "acquisition tails" for
  the browser database timescales were sometimes not enabled or disabled
  when acquisition was en-/disabled or moved from/to the browser database.

See file CHANGES.RCS for a detailed log of revisions of all files.

+++ epgctl/epgctxctl.c
@@ -31,7 +31,7 @@
  *
  *  Author: Tom Zoerner
  *
- *  $Id: epgctxctl.c,v 1.19 2002/03/22 21:13:56 tom Exp tom $
+ *  $Id: epgctxctl.c,v 1.20 2002/05/11 15:44:51 tom Exp tom $
  */
 
 #define DEBUG_SWITCH DEBUG_SWITCH_EPGCTL
@@ -854,7 +854,7 @@ time_t EpgContextCtl_GetAiUpdateTime( uint cni )
 // - used by EPG scan for already known providers (the provider might have been
 //   found in passive mode when the channel freq. could not be determined)
 //
-bool EpgContextCtl_UpdateFreq( uint cni, ulong freq )
+bool EpgContextCtl_UpdateFreq( uint cni, uint freq )
 {
    CTX_CACHE * pContext;
    bool result;
@@ -864,7 +864,7 @@ bool EpgContextCtl_UpdateFreq( uint cni, ulong freq )
    pContext = EpgContextCtl_SearchCni(cni);
    if ((pContext != NULL) && (pContext->state == CTX_CACHE_OPEN))
    {  // db is open -> just set the frequency there
-      dprintf3("EpgContextCtl-UpdateFreq: update freq for opened db 0x%04X: %lu -> %lu\n", cni, pContext->pDbContext->tunerFreq, freq);
+      dprintf3("EpgContextCtl-UpdateFreq: update freq for opened db 0x%04X: %u -> %u\n", cni, pContext->pDbContext->tunerFreq, freq);
 
       // set the flag that the db needs to be saved
       if (pContext->pDbContext->tunerFreq != freq)
@@ -875,7 +875,7 @@ bool EpgContextCtl_UpdateFreq( uint cni, ulong freq )
    }
    else
    {  // db is not open -> write the frequency into the file header
-      debug2("EpgContextCtl-UpdateFreq: update freq for non-open db 0x%04X: %lu", cni, freq);
+      debug2("EpgContextCtl-UpdateFreq: update freq for non-open db 0x%04X: %u", cni, freq);
       result = EpgDbDumpUpdateHeader(cni, freq);
    }
    return result;
@@ -956,11 +956,11 @@ const uint * EpgContextCtl_GetProvList( uint * pCount )
 // - for databases which are not cached because they are defective or have and
 //   incompatible version, the tuner frequency is read from the header
 //
-uint EpgContextCtl_GetFreqList( uint ** ppProvList, ulong ** ppFreqList )
+uint EpgContextCtl_GetFreqList( uint ** ppProvList, uint ** ppFreqList )
 {
    CTX_CACHE * pContext;
    uint maxCount, idx;
-   ulong freq;
+   uint freq;
 
    // determine the number of providers, including "error" contexts
    maxCount = 0;
